requirements.txt
basicsr==1.4.2
deep-sort-realtime==1.3.2
cython<3.0.0
gitpython>=3.1.30
jupyter==1.0.0
onnxruntime
opencv-python 
matplotlib 
moviepy==1.0.3
mss
pandas
pillow 
plotly
pyyaml==5.4.1
realesrgan==0.3.0
supervision==0.19.0
torchvision 
transformers
utlralytics==8.1.43

# esgan 
numpy
torch>=1.7
tqdm

# hat 
# basicsr == 1.3.4.9

# # may need to run following contraints to install detectron2
# !echo "cython<3" > /tmp/constraint.txt
# !PIP_CONSTRAINT=/tmp/constraint.txt pip install 
# 'git+https://github.com/facebookresearch/detectron2.git'

# mbari specific apia
fathomnet